backend/requirements.txt
Flask>=3.0
flask-cors>=4.0
openai>=1.10
python-dotenv>=1.0